Skip over CPUs in the idle loop waiting for SCHED.
Skipping CPUs with the CPU$V_SCHED bit set appears to provide a
more accurate load average count than before.

OpenVMS Alpha Internals: Scheduling & Process Control describes the
idle loop in 2.7.3, including step 9, when the code sees that the idle
bit has been cleared for its CPU ID, and it sets CPU$V_SCHED in the
CPU database to indicate that it's attempting to acquire SCHED.

My theory is that my timer queue callback is sampling states when one
or more CPUs are in the process of rescheduling themselves out of the
idle state, but unable to acquire the SCHED spinlock because they've
been interrupted by the timer callback and my own acquiring of SCHED.

This doesn't entirely explain why a CPU could have a KTB pointing to
a kernel thread that's also in one of the computable or wait queues,
and I suspect that the scheduler in current versions of VMS operates
differently from the description of OpenVMS V7.0 in the internals book.